+++ IDesign/IDesign.Extension/IDesign.ExtensionPackage.cs
@@ -2,7 +2,10 @@
 using System.Runtime.InteropServices;
 using System.Threading;
 using System.Threading.Tasks;
+using EnvDTE;
+using System.ComponentModel;
 using Microsoft.VisualStudio.Shell;
+using Microsoft.VisualStudio.Shell.Interop;
 
 namespace IDesign.Extension
 {
@@ -24,8 +27,11 @@ namespace IDesign.Extension
     ///         &gt; in .vsixmanifest file.
     ///     </para>
     /// </remarks>
+    [ProvideAutoLoad(UIContextGuids80.SolutionExists, PackageAutoLoadFlags.BackgroundLoad)]
     [PackageRegistration(UseManagedResourcesOnly = true, AllowsBackgroundLoading = true)]
     [Guid(PackageGuidString)]
+    [ProvideOptionPage(typeof(IDesignOptionPageGrid),
+    "IDesign", "Privacy", 0, 0, true)]
     public sealed class IDesignExtensionPackage : AsyncPackage
     {
         /// <summary>
@@ -54,8 +60,23 @@ namespace IDesign.Extension
             // When initialized asynchronously, the current thread may be a background thread at this point.
             // Do any initialization that requires the UI thread after switching to the UI thread.
             await JoinableTaskFactory.SwitchToMainThreadAsync(cancellationToken);
+            var dte = (DTE)GetService(typeof(DTE));
+            SubscribeBuildEvents.Initialize(dte);
         }
 
         #endregion
+    }
+    public class IDesignOptionPageGrid : DialogPage
+    {
+        private bool _doLogData = false;
+
+        [Category("Privacy")]
+        [DisplayName("Log data")] 
+        [Description("Whether IDesign can log your data. The data which gets logged are your actions and your source code. This is used for research. This option is turned off by default.")]
+        public bool DoLogData
+        {
+            get { return _doLogData; }
+            set { _doLogData = value; }
+        }
     }
 }