+++ IDesign/IDesign.Extension/Views/HomeView.xaml.cs
@@ -1,5 +1,9 @@
-﻿using System.Windows;
+﻿using System;
+using System.Windows;
 using System.Windows.Controls;
+using Microsoft.VisualStudio;
+using Microsoft.VisualStudio.Shell;
+using Microsoft.VisualStudio.Shell.Interop;
 
 namespace IDesign.Extension.Views
 {
@@ -16,6 +20,17 @@ namespace IDesign.Extension.Views
         private void HomeView_OnSizeChanged(object sender, SizeChangedEventArgs e)
         {
             ButtonsPanel.Orientation = e.NewSize.Width >= ButtonsPanel.Children.Count*190 ? Orientation.Horizontal : Orientation.Vertical;
+        }
+        public void OnSettingsClickEvent(object sender, RoutedEventArgs e)
+        {
+            ThreadHelper.ThrowIfNotOnUIThread();
+            Guid guid = typeof(IDesignExtensionPackage).GUID;
+            var vsShell = (IVsShell)ServiceProvider.GlobalProvider.GetService(typeof(IVsShell));
+            if (vsShell.IsPackageLoaded(ref guid, out var myPackage)
+               == Microsoft.VisualStudio.VSConstants.S_OK)
+            {
+                ((IDesignExtensionPackage)myPackage).ShowOptionPage(typeof(IDesignOptionPageGrid));
+            }
         }
     }
 }